7.19 completed(?)
Question mark is due to me not being sure if i squash all the bugs and completed all the parts. I *think* i got everything, but if you're reading this and i missed something feel free to let me know so i can fix it. or mark me points for it, whichever is easier on your part.